Add axios plus axios default configuration and updated the currentDailyWeatherContainer
