+++ package.json
@@ -12,6 +12,7 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.53",
     "@types/react-dom": "^16.9.8",
+    "axios": "^0.21.0",
     "date-fns": "^2.16.1",
     "normalize.css": "^8.0.1",
     "react": "^17.0.1",
